Exception helper must use privileged SuppressedMuleException
